backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via jupyter-server
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.2.3
    # via isoduration
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
attrs==22.2.0
    # via jsonschema
autopep8==1.5.7
    # via -r requirements.in
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.2
    # via nbconvert
bleach==6.0.0
    # via nbconvert
boto3==1.17.109
    # via -r requirements.in
botocore==1.20.112
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via argon2-cffi-bindings
charset-normalizer==2.1.1
    # via requests
colorama==0.4.6
    # via ipython
comm==0.1.2
    # via ipykernel
coverage==7.1.0
    # via -r requirements.in
dataclasses==0.6
    # via -r requirements.in
debugpy==1.6.6
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
django==3.2.6
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.7.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.0
    # via -r requirements.in
executing==1.2.0
    # via stack-data
faker==8.1.3
    # via -r requirements.in
fastjsonschema==2.16.2
    # via nbformat
fqdn==1.5.1
    # via jsonschema
graphviz==0.20.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   jsonschema
    #   requests
importlib-metadata==6.0.0
    # via
    #   jupyter-client
    #   nbconvert
ipykernel==6.21.2
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   nbclassic
    #   notebook
    #   qtconsole
ipython==8.10.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipython-genutils==0.2.0
    # via
    #   nbclassic
    #   notebook
    #   qtconsole
ipywidgets==8.0.4
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   nbclassic
    #   nbconvert
    #   notebook
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonpointer==2.3
    # via jsonschema
jsonschema[format-nongpl]==4.17.3
    # via
    #   jupyter-events
    #   nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==8.0.3
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.5.1
    # via jupyter
jupyter-core==5.2.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter-events==0.6.3
    # via jupyter-server
jupyter-server==2.3.0
    # via
    #   nbclassic
    #   notebook-shim
jupyter-server-terminals==0.4.4
    # via jupyter-server
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-widgets==3.0.5
    # via ipywidgets
markupsafe==2.1.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
mistune==2.0.5
    # via nbconvert
nbclassic==0.5.1
    # via notebook
nbclient==0.7.2
    # via nbconvert
nbconvert==7.2.9
    # via
    #   jupyter
    #   jupyter-server
    #   nbclassic
    #   notebook
nbformat==5.7.3
    # via
    #   jupyter-server
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.6
    # via
    #   ipykernel
    #   nbclassic
    #   notebook
notebook==6.5.2
    # via jupyter
notebook-shim==0.2.2
    # via nbclassic
numpy==1.24.2
    # via pandas
packaging==23.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbconvert
    #   qtpy
pandas==1.2.3
    # via -r requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pdfkit==1.0.0
    # via -r requirements.in
pickleshare==0.7.5
    # via ipython
pillow==9.4.0
    # via -r requirements.in
platformdirs==3.0.0
    # via jupyter-core
prometheus-client==0.16.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
prompt-toolkit==3.0.36
    # via
    #   ipython
    #   jupyter-console
psutil==5.9.4
    # via ipykernel
psycopg2-binary==2.8.6
    # via -r requirements.in
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via autopep8
pycparser==2.21
    # via cffi
pycryptodome==3.17
    # via -r requirements.in
pydot==1.4.2
    # via -r requirements.in
pydot-ng==2.0.0
    # via -r requirements.in
pygments==2.14.0
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
    #   qtconsole
pyjwt==2.6.0
    # via djangorestframework-simplejwt
pyparsing==3.0.9
    # via
    #   pydot
    #   pydot-ng
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   arrow
    #   botocore
    #   faker
    #   jupyter-client
    #   pandas
python-json-logger==2.0.6
    # via jupyter-events
pytz==2022.7.1
    # via
    #   django
    #   pandas
pywin32==305
    # via jupyter-core
pywinpty==2.0.10
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   terminado
pyyaml==6.0
    # via jupyter-events
pyzmq==25.0.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   nbclassic
    #   notebook
    #   qtconsole
qtconsole==5.4.0
    # via jupyter
qtpy==2.3.0
    # via qtconsole
requests==2.28.1
    # via -r requirements.in
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
s3transfer==0.4.2
    # via boto3
send2trash==1.8.0
    # via
    #   jupyter-server
    #   nbclassic
    #   notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.0
    # via anyio
soupsieve==2.4
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
stack-data==0.6.2
    # via ipython
terminado==0.17.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
    #   nbclassic
    #   notebook
text-unidecode==1.3
    # via faker
tinycss2==1.2.1
    # via nbconvert
toml==0.10.2
    # via autopep8
tornado==6.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   nbclassic
    #   notebook
    #   terminado
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   matplotlib-inline
    #   nbclassic
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
uri-template==1.2.0
    # via jsonschema
urllib3==1.26.14
    # via
    #   botocore
    #   requests
wcwidth==0.2.6
    # via prompt-toolkit
webcolors==1.12
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.5.1
    # via jupyter-server
widgetsnbextension==4.0.5
    # via ipywidgets
zipp==3.14.0
    # via importlib-metadata
